save step appended. model logics improved

+++ displaymeltmodel.cpp
@@ -3,13 +3,14 @@
 #include "ui_displaymeltmodel.h"
 #include "meltmodel.h"
 
+#include <QGraphicsView>
+
 DisplayMeltmodel::DisplayMeltmodel(QWidget *parent) :
     QMainWindow(parent),
     ui(new Ui::displaymeltmodel),
     m_meltmodel(NULL)
 {
     ui->setupUi(this);
-    //connect(ui->start, SI)
     connect(ui->start, SIGNAL(clicked()), this, SLOT(initModel()));
 }
 
@@ -20,7 +21,7 @@ DisplayMeltmodel::~DisplayMeltmodel()
 
 void DisplayMeltmodel::initModel()
 {
-    startNewModel(ui->x->text().toInt(), ui->y->text().toInt(), ui->temperature->text().toInt());
+    startNewModel(ui->x->text().toInt(), ui->y->text().toInt(), ui->temperature->text().toDouble());
 }
 
 void DisplayMeltmodel::step()
@@ -28,7 +29,7 @@ void DisplayMeltmodel::step()
     m_meltmodel->processStep();
 }
 
-void DisplayMeltmodel::startNewModel(int _width, int _height, int _startTemperature)
+void DisplayMeltmodel::startNewModel(int _width, int _height, double _startTemperature)
 {
     delete m_meltmodel;
     m_meltmodel = new MeltModel(_width, _height, _startTemperature);
